+++ core/src/main/java/org/mwg/plugin/NodeStateCallback.java
@@ -3,6 +3,6 @@ package org.mwg.plugin;
 @FunctionalInterface
 public interface NodeStateCallback {
 
-    void on(long attributeKey, byte elemType, Object elem);
+    void on(int attributeKey, byte elemType, Object elem);
 
 }
