+++ core/src/main/java/org/mwg/struct/ENode.java
@@ -6,15 +6,15 @@ public interface ENode {
 
     ENode set(String name, byte type, Object value);
 
-    ENode setAt(long key, byte type, Object value);
+    ENode setAt(int key, byte type, Object value);
 
     Object get(String name);
 
-    Object getAt(long key);
+    Object getAt(int key);
 
     Object getOrCreate(final String key, final byte type);
 
-    Object getOrCreateAt(final long key, final byte type);
+    Object getOrCreateAt(final int key, final byte type);
 
     void drop();
